fix: staff area ui mixins and removing the old one
