+++ openassessment/xblock/ui_mixins/legacy/handlers_mixin.py
@@ -5,8 +5,11 @@ JSON handlers for the old-style ORA UI
 
 from xblock.core import XBlock
 import logging
+
+from openassessment.xblock.staff_area_mixin import require_course_staff
 from openassessment.xblock.ui_mixins.legacy.peer_assessments.actions import peer_assess
 from openassessment.xblock.ui_mixins.legacy.self_assessments.actions import self_assess
+from openassessment.xblock.ui_mixins.legacy.staff_assessments.actions import staff_assess
 from openassessment.xblock.ui_mixins.legacy.student_training.actions import training_assess
 from openassessment.xblock.ui_mixins.legacy.submissions.actions import (
     submit,
@@ -71,8 +74,13 @@ class LegacyHandlersMixin:
 
     @XBlock.json_handler
     def training_assess(self, data, suffix=""):
-        logger.warning("block.training_assess")
         return training_assess(self.api_data, data)
+
+    @XBlock.json_handler
+    @require_course_staff("STUDENT_INFO")
+    @verify_assessment_parameters
+    def staff_assess(self, data, suffix=""):
+        return staff_assess(self.api_data, data)
 
     # Utils
